+++ analysis.go
@@ -35,6 +35,7 @@ const (
 	AnalysisModeDNSReportCorrelation    AnalysisMode = "dns_report_correlation"
 	AnalysisModeThreatCandidates        AnalysisMode = "threat_candidates"
 	AnalysisModeSourceEnrichment        AnalysisMode = "source_enrichment"
+	AnalysisModeJurisdictionContext     AnalysisMode = "jurisdiction_context"
 	AnalysisModeCampaignValidation      AnalysisMode = "campaign_configuration_validation"
 	AnalysisModeCampaignClassification  AnalysisMode = "campaign_classification"
 )